Fix tests and update node version
