+++ .github/workflows/learning-web-release.yaml
@@ -21,7 +21,7 @@ jobs:
       - name: Setup node.js
         uses: actions/setup-node@v1
         with:
-          node-version: "13.11"
+          node-version: "14.17"
 
       - name: Install modules
         working-directory: learning/web
