+++ learning/web/config/__mocks__/gatsby.js
@@ -1,6 +1,5 @@
 const React = require("react")
 const gatsby = jest.requireActual("gatsby")
-
 module.exports = {
   ...gatsby,
   graphql: jest.fn(),
