--- learning/web/config/enzyme.js
@@ -1,4 +1,0 @@
-const enzyme = require("enzyme")
-const Adapter = require("enzyme-adapter-react-16")
-
-enzyme.configure({ adapter: new Adapter() })
