+++ learning/web/config/jest-preprocess.js
@@ -1,5 +1,4 @@
 const babelOptions = {
   presets: ["babel-preset-gatsby"],
 }
-
-module.exports = require("babel-jest").createTransformer(babelOptions)
+module.exports = require("babel-jest").default.createTransformer(babelOptions)
