+++ learning/web/package.json
@@ -47,14 +47,14 @@
     "uuid": "^8.3.2"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.17.12",
+    "@babel/preset-react": "^7.17.12",
     "babel-jest": "^28.1.0",
     "babel-preset-gatsby": "^2.15.0",
     "cypress": "^9.7.0",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.6",
-    "enzyme-to-json": "^3.6.2",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^28.1.0",
+    "jest-environment-jsdom": "^28.1.0",
     "minimatch": "^5.1.0",
     "pa11y-ci": "^3.0.1",
     "prettier": "^2.6.2",
